Make acceleration and collisions work on POJOs
POJO = plain old JavaScript objects. This will remove the need
for PhysicsObject and PhysicsGroup types, and make the system
more declarative overall. What I'm trying to accomplish is making
the physics similar to the drawing, where you hand it the state
each tick and have it figure out the best way to update all the
physics for the objects in the state